test: 리프레시 토큰을 post 통신하여 액세스 토큰, 리프레시 토큰을 반환하는 함수 테스트
