+++ frontend/src/mocks/mockData/token.ts
@@ -1,0 +1,5 @@
+export const MOCK_TOKEN = {
+  accessToken:
+    'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJtZW1iZXJJZCI6NiwiaWF0IjoxNjkzODM2NDgzLCJleHAiOjE2OTM5MjI4ODN9.iEnSoZVT-PLPjFVaDvMRR9dmCWjbARyXm_giEkDFoTw',
+  refreshToken: 'refreshToken',
+};
